Edits from rev2->master pull request
