+++ src/Views/Arrow/ArrowSpriteFactory.java
@@ -1,0 +1,4 @@
+package Views.Arrow;
+
+public class ArrowSpriteFactory {
+}
