+++ src/Views/Environment/Background/BackgroundSpriteFactory.java
@@ -1,0 +1,4 @@
+package Views.Background;
+
+public class BackgroundSpriteFactory {
+}
