+++ src/Views/Environment/EndscreenDrawer.java
@@ -1,0 +1,4 @@
+package Views.Environment.Endscreen;
+
+public class EndscreenDrawer {
+}
