+++ src/Views/GameObject/Ball/BallDrawer.java
@@ -1,0 +1,4 @@
+package Views.GameObject.Ball;
+
+public class BallDrawer {
+}
