+++ src/Views/GameObject/Ball/BallSpriteFactory.java
@@ -1,0 +1,4 @@
+package Views.GameObject.Ball;
+
+public class BallSpriteFactory {
+}
