+++ src/Views/GameObject/GameObjectDrawer.java
@@ -1,0 +1,4 @@
+package Views.GameObject;
+
+public class GameObjectDrawer {
+}
