+++ src/Views/GameObject/Goal/GoalDrawer.java
@@ -1,0 +1,4 @@
+package Views.GameObject.Goal;
+
+public class GoalDrawer {
+}
