+++ src/Views/GameObject/Goal/GoalSpriteFactory.java
@@ -1,0 +1,4 @@
+package Views.GameObject.Goal;
+
+public class GoalSpriteFactory {
+}
